+++ src/song/cmarker.h
@@ -11,7 +11,7 @@ class CMarker
     : public QObject
 {
     Q_OBJECT
-    Q_PROPERTY(int     measure READ measure WRITE measure NOTIFY measureChanged)
+    Q_PROPERTY(qint16  measure READ measure WRITE measure NOTIFY measureChanged)
     Q_PROPERTY(QString title READ title WRITE title NOTIFY titleChanged)
     Q_PROPERTY(QColor  color READ color WRITE color NOTIFY colorChanged)
 
@@ -23,11 +23,11 @@ public:
     inline bool operator==(const CMarker &obj) const { return measure() == obj.measure() && title() == obj.title() && color() == obj.color(); }
 
     // ---API--- //
-    inline int            measure() const { return m_measure; }
+    inline qint16         measure() const { return m_measure; }
     inline const QString& title()   const { return m_title; }
     inline const QColor&  color()   const { return m_color; }
 
-    inline void measure(const int val)    { m_measure = val; emit measureChanged(); }
+    inline void measure(const qint16 val) { m_measure = val; emit measureChanged(); }
     inline void title(const QString &val) { m_title = val; emit titleChanged(); }
     inline void color(const QColor &val)  { m_color = val; emit colorChanged(); }
 
@@ -40,7 +40,7 @@ private:
     static const Qt::GlobalColor DEFAULT_COLOR = Qt::red;
     static const char*           DEFAULT_TITLE;
 
-    int     m_measure;
+    qint16  m_measure;
     QString m_title;
     QColor  m_color;
 };
